backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Backend logic for Content Visualisation project"
authors = [
  "William Simson van Dijkhuizen <psyws2@nottingham.ac.uk>",
  "Oluwagbemiga Ayanola <psyoa5@nottingham.ac.uk>",
  "Omar Zeghoauni <psyoz1@nottingham.ac.uk>",
  "Robert Sykes <psyrs13@nottingham.ac.uk>",
  "Sean Butcher <psysb11@nottingham.ac.uk>",
  "Emre Ozsaruhan <psyeo1@nottingham.ac.uk>"
]
include = ["debug_server.py", "README.md"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
PyMuPDF = "^1.19.3"
fastapi = "^0.70.1"
numpy = "^1.21.5"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
python-multipart = "^0.0.5"
spacy = "^3.2.1"
en-core-web-sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz"}
gunicorn = "^20.1.0"
isort = {extras = ["colors"], version = "^5.10.1"}
black = "^21.12b0"
httpx = "^0.22.0"
bs4 = "^0.0.1"
lxml = "^4.8.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "^21.11b0"
pytest = "^6.2.5"
isort = "^5.10.1"
pydocstyle = "^6.1.1"
coverage = "^6.1.2"
flake8 = "^4.0.1"
flake8-docstrings = "^1.6.0"
pep8-naming = "^0.12.1"
flake8-blind-except = "^0.2.0"
flake8-builtins = "^1.5.3"
flake8-black = "^0.2.3"
Sphinx = "^4.3.1"
colorama = "^0.4.4"

[tool.isort]
profile = "black"
extend_skip = ["docs"]

[tool.pydocstyle]
convention = "google"
match = "(?!__init).*\\.py"
match-dir = "^(?!(.venv|docs)).*"

[tool.black]
target-version = ["py310"]
extend-exclude = "docs"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
